SQUASH: Refactor importing stuff
